+++ lib/keen_auth_permissions/database/parsers/update_translation_processor.ex
@@ -28,7 +28,7 @@ defmodule KeenAuthPermissions.Database.Parsers.UpdateTranslationParser do
     {:ok, results}
   end
 
-  defp parse_row([created_at, created_by, updated_at, updated_by, translation_id, language_code, tenant_id, data_group, data_object_code, data_object_id, value, ua_search_data, ts_search_data]) do
+  defp parse_row([created_at, created_by, updated_at, updated_by, translation_id, language_code, tenant_id, data_group, data_object_code, data_object_id, value, nrm_search_data, ts_search_data]) do
     {:ok, %UpdateTranslationModel{
       created_at: created_at,
       created_by: created_by,
@@ -41,7 +41,7 @@ defmodule KeenAuthPermissions.Database.Parsers.UpdateTranslationParser do
       data_object_code: data_object_code,
       data_object_id: data_object_id,
       value: value,
-      ua_search_data: ua_search_data,
+      nrm_search_data: nrm_search_data,
       ts_search_data: ts_search_data,
     }}
   end